moved vue_multi_component to vue example
