+++ examples/vue/package.json
@@ -21,7 +21,13 @@
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
     "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.11"
+    "vue-template-compiler": "^2.6.11",
+    "vue-loader": "15.9.7",
+    "vue-style-loader": "4.1.3",
+    "css-loader": "5.2.6",
+    "babel-loader": "8.2.2",
+    "webpack": "^4.0.0",
+    "webpack-cli": "4.7.2"
   },
   "eslintConfig": {
     "root": true,
