+++ examples/vue/src/App.vue
@@ -6,7 +6,8 @@
 </template>
 
 <script>
-import HelloWorld from './components/HelloWorld.vue'
+
+import HelloWorld from 'HelloWorld'
 
 export default {
   name: 'App',
