+++ examples/vue/src/components/HelloWorld/BUILD
