+++ examples/vue/src/components/HelloWorld/HelloWorld.vue
