+++ examples/vue/src/components/HelloWorld/package.json
