+++ examples/vue/src/components/HelloWorld/speedcheck.sh
