+++ examples/vue/src/components/HelloWorld/webpack.config.js
