--- examples/vue_multi_component/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "hello-world",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "serve": "bazel run :serve",
-    "build": "bazel build :build",
-    "test": "bazel test //...",
-    "lint": "vue-cli-service lint"
-  },
-  "dependencies": {
-    "core-js": "^3.6.5",
-    "vue": "^2.6.11"
-  },
-  "devDependencies": {
-    "@bazel/bazelisk": "^1.7.2",
-    "@bazel/ibazel": "^0.13.2",
-    "@vue/cli-plugin-babel": "~4.4.0",
-    "@vue/cli-plugin-eslint": "~4.4.0",
-    "@vue/cli-service": "~4.4.0",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.11",
-    "vue-loader": "15.9.7",
-    "vue-style-loader": "4.1.3",
-    "css-loader": "5.2.6",
-    "babel-loader": "8.2.2",
-    "webpack": "^4.0.0",
-    "webpack-cli": "4.7.2"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended"
-    ],
-    "parserOptions": {
-      "parser": "babel-eslint"
-    },
-    "rules": {}
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ]
-}
