--- examples/vue_multi_component/smoke.test.js
@@ -1,6 +1,0 @@
-const assert = require('assert');
-const fs = require('fs');
-const path = require('path');
-
-process.chdir(path.join(__dirname, 'dist'));
-assert.ok(fs.readFileSync('index.html', 'utf-8').match(/href=\/js\/app/));
