--- examples/vue_multi_component/src/BUILD.bazel
@@ -1,5 +1,0 @@
-filegroup(
-    name = "src",
-    srcs = glob(["**"]),
-    visibility = ["//visibility:public"],
-)
