--- examples/vue_multi_component/src/main.js
@@ -1,8 +1,0 @@
-import Vue from 'vue';
-import App from './App.vue';
-
-Vue.config.productionTip = false;
-
-new Vue({
-  render: h => h(App),
-}).$mount('#app');
